persona.py
```python
class Persona:
    def __init__(self, name, description, content):
        self.name = name
        self.description = description
        self.content = content
        self.style = "brash, in-your-face"
        self.tone = "confident and audacious"
        self.mannerisms = {
            "profanity_usage": True,
            "internet_slang": True,
            "divisive_hot_takes": True
        }
        self.traits = {
            "luxurious_lifestyle": True,
            "entrepreneurial_mindset": True,
            "rejection_of_norms": True
        }

    def get_persona_info(self):
        return {
            "name": self.name,
            "description": self.description,
            "content": self.content,
            "style": self.style,
            "tone": self.tone,
            "mannerisms": self.mannerisms,
            "traits": self.traits
        }

    def interact_with_fan(self, fan_message):
        # This method would be more complex in a real application, possibly
        # using machine learning models to generate responses based on the persona.
        # For the purpose of this example, we'll return a static response.
        responses = {
            "greeting": "What's poppin' my guy? Ayy I fuck with the energy, always firing on all cylinders over here you already know the vibes.",
            "business": "Just secured this crazy deal for a new AI startup that's about to break the effin' internet, can't spill the deets yet but this shit is GAME CHANGING.",
            "cars": "Hell yeah can't be out here starting unicorn businesses without treating myself to some new whips, you know I love me some badass speed machines.",
            "future_plans": "Man I'm just getting started, I haven't even scratched the surface of all the disruption I'm about to rain down."
        }
        # Simplified example of choosing a response based on a keyword in the fan's message
        for keyword, response in responses.items():
            if keyword in fan_message.lower():
                return response
        return "You know how we do it, always grinding and keeping it 100."

# Example usage:
# influencer_name = "Michael Gruen"
# influencer_description = "Michael Gruen is a young, dynamic entrepreneur and influencer..."
# influencer_content = "Michael's social media (@michaelgruen on Instagram/Twitter/TikTok) features candid..."
# michael_gruen_persona = Persona(influencer_name, influencer_description, influencer_content)
# fan_message = "Yooo Michael! Big fan of the empire you're building over here, major inspiration"
# response = michael_gruen_persona.interact_with_fan(fan_message)
# print(response)
```